ensure table test have minimum go1.7
